+++ allauth/__init__.py
@@ -8,7 +8,7 @@ r"""
 
 """
 
-VERSION = (0, 41, 0, 'final', 0)
+VERSION = (0, 42, 0, 'final', 0)
 
 __title__ = 'django-allauth'
 __version_info__ = VERSION
@@ -16,4 +16,4 @@ __version__ = '.'.join(map(str, VERSION[:3])) + ('-{}{}'.format(
     VERSION[3], VERSION[4] or '') if VERSION[3] != 'final' else '')
 __author__ = 'Raymond Penners'
 __license__ = 'MIT'
-__copyright__ = 'Copyright 2010-2019 Raymond Penners and contributors'
+__copyright__ = 'Copyright 2010-2020 Raymond Penners and contributors'