+++ allauth/socialaccount/providers/instagram/provider.py
@@ -9,9 +9,6 @@ class InstagramAccount(ProviderAccount):
     def get_profile_url(self):
         return self.PROFILE_URL + self.account.extra_data.get('username')
 
-    def get_avatar_url(self):
-        return self.account.extra_data.get('profile_picture')
-
     def to_str(self):
         dflt = super(InstagramAccount, self).to_str()
         return self.account.extra_data.get('username', dflt)
@@ -23,17 +20,16 @@ class InstagramProvider(OAuth2Provider):
     account_class = InstagramAccount
 
     def extract_extra_data(self, data):
-        return data.get('data', {})
+        return data
 
     def get_default_scope(self):
         return ['basic']
 
     def extract_uid(self, data):
-        return str(data['data']['id'])
+        return str(data['id'])
 
     def extract_common_fields(self, data):
-        return dict(username=data['data'].get('username'),
-                    name=data['data'].get('full_name'))
+        return dict(username=data.get('username'))
 
 
 provider_classes = [InstagramProvider]