+++ ChangeLog.rst
@@ -1,13 +1,15 @@
-0.42.0 (Unreleased)
+0.42.0 (2020-05-24)
 *******************
 
 Note worthy changes
 -------------------
 
 - New provider: Apple.
-- New provider: Yandex (OAuth2)
+- New providers: EDX, Yandex, Mixer.
 - Fixed Twitch ``get_avatar_url()`` method to use the profile picture retrieved
   by new user details endpoint introduced in version 0.40.0.
+
+- The Facebook API version now defaults to v7.0.
 
 
 0.41.0 (2019-12-18)
@@ -27,6 +29,8 @@ Note worthy changes
 - You can now store OAuth credentials directly in your
   ``settings.SOCIALACCOUNT_PROVIDERS`` settings instead of storing them in the
   database using a ``SocialApp`` record.
+
+- Adding Keycloak Provider
 
 
 Backwards incompatible changes